requirements.txt
anyascii==0.3.2
asgiref==3.8.1
beautifulsoup4==4.13.4
certifi==2025.4.26
charset-normalizer==3.4.2
coverage==7.8.1
defusedxml==0.7.1
Django==5.2.1
django-cors-headers==4.7.0
django-environ==0.12.0
django-filter==25.1
django-modelcluster==6.4
django-permissionedforms==0.1
django-prometheus==2.3.1
django-stubs-ext==5.2.0
django-taggit==6.1.0
django-tasks==0.6.1
django-treebeard==4.7.1
djangorestframework==3.16.0
draftjs_exporter==5.1.0
et_xmlfile==2.0.0
filetype==1.2.0
gunicorn==23.0.0
idna==3.10
iniconfig==2.1.0
laces==0.1.2
openpyxl==3.1.5
packaging==25.0
pillow==11.2.1
pillow_heif==0.21.0
pluggy==1.6.0
prometheus_client==0.22.0
psycopg2-binary==2.9.10
pytest==8.3.5
pytest-django==4.11.1
python-decouple==3.8
requests==2.32.3
soupsieve==2.7
sqlparse==0.5.3
telepath==0.3.1
typing_extensions==4.13.2
urllib3==2.4.0
wagtail==6.4.1
wagtail-cache==2.5.1
wagtail-modeladmin==2.2.0
wagtail-seo==3.0.0
whitenoise==6.9.0
Willow==1.10.0
prometheus-client
django-filter
